@@ -1,2 +1,6 @@
-package maxnazarati.web.pattern.crud.foundation;public class Model<ID> {
+package maxnazarati.web.pattern.crud.foundation;
+
+public interface Model<ID> {
+    ID id();
+    void updateBy(Model<ID> other);
 }
