@@ -1,2 +1,8 @@
-package maxnazarati.web.pattern.crud.foundation;public class Query<ID> {
+package maxnazarati.web.pattern.crud.foundation;
+
+import java.util.List;
+
+public interface Query<ID> {
+    List<ID> ids();
+
 }
