@@ -1,2 +1,7 @@
-package maxnazarati.web.pattern.crud.foundation;public interface Validator<M> {
+package maxnazarati.web.pattern.crud.foundation;
+
+public interface Validator<M> {
+    public void validateCreationOrUpdate(M model);
+    public void validateDeletion(M model);
+
 }
